# apps/core/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoreX Service API",
    "description": "HTTP 状态码始终为 200；业务结果见响应体 body.code（200000=成功）。错误码规则见 /guide/error-codes.md。Exception.details 字段仅在开发环境返回。鉴权：JWT Bearer；Apifox 环境变量统一使用 {{token}}（来自 signin.data.token）。密码登录 / 注册 / 发 OTP 前须先 POST /api/v1/auth/captcha 获取 captchaKey（go-captcha 行为验证码）。",
    "contact": {
      "name": "CoreX Team",
      "email": "15638470820@163.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "development"
    },
    {
      "url": "https://api.example.com",
      "description": "production"
    }
  ],
  "paths": {
...
    "/api/v1/gateway/providers": {
      "get": {
        "tags": [
          "Gateway"
        ],
        "summary": "供应商列表",
        "operationId": "gateway.providers",
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Gateway"
        ],
        "summary": "创建供应商",
        "operationId": "gateway.providerWrite",
        "requestBody": {
          "description": "供应商信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderWriteP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/sso/connections": {
      "get": {
        "tags": [
          "SSO"
        ],
        "summary": "SSO 连接列表",
        "operationId": "sso.connections",
        "responses": {
          "200": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "SSO"
        ],
        "summary": "创建 SSO 连接",
        "operationId": "sso.connectionWrite",
        "requestBody": {
          "description": "连接信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoConnectionWriteP"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/sso/{id}/authorize": {
      "get": {
        "tags": [
          "SSO"
        ],
        "summary": "发起 OIDC 授权（302 到 IdP）",
        "operationId": "sso.authorize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "连接 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "连接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          },
          "302": {
            "description": "重定向到身份提供商"
          }
        }
      }
    },
    "/api/v1/sso/{id}/callback": {
      "get": {
        "tags": [
          "SSO"
        ],
        "summary": "OIDC 回调（换取 token）",
        "operationId": "sso.callback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "连接 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "授权码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exception": {
        "type": "object",
        "description": "异常响应结构",
        "required": [
          "code",
          "success",
          "msg",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "业务错误码 (非0表示各种业务错误)"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "详细错误信息（仅非生产环境返回）"
          },
          "msg": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProviderWriteP": {
        "type": "object",
        "required": [
          "kind",
          "name",
          "baseUrl"
        ],
        "properties": {
          "apiKey": {
            "type": [
              "string",
              "null"
            ],
            "description": "明文 API Key；入库前 AES 加密"
          },
          "baseUrl": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "description": "openai / anthropic / deepseek / qwen / zhipu / ollama"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "description": "ACTIVE / DISABLED"
          }
        }
      },
...
      "SsoConnectionWriteP": {
        "type": "object",
        "required": [
          "tenantID",
          "provider",
          "issuer",
          "clientID",
          "redirectUri"
        ],
        "properties": {
          "clientID": {
            "type": "string"
          },
          "clientSecret": {
            "type": [
              "string",
              "null"
            ],
            "description": "明文 client secret；入库前 AES 加密"
          },
          "issuer": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "description": "oidc / saml（saml 预留）"
          },
          "redirectUri": {
            "type": "string"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "tenantID": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录 `POST /api/v1/auth/signin` 返回的 `data.token`。\n\n请求头：`Authorization: Bearer <token>`。\n\n**Apifox**：鉴权组件 Token 请填 `{{token}}`（勿用 `{{bearerToken}}`）；环境变量名统一为 `token`，可在登录接口后置操作写入。",
        "x-default": "{{token}}",
        "x-apifox-default": "{{token}}"
      }
    }
  },
...
}